@@ -1,16 +1,20 @@
 module github.com/kujenga/website/exp/go-templates
 
-go 1.16
+go 1.21
+
+require (
+	github.com/Masterminds/sprig v2.22.0+incompatible
+	gopkg.in/yaml.v3 v3.0.0
+)
 
 require (
 	github.com/Masterminds/goutils v1.1.1 // indirect
 	github.com/Masterminds/semver v1.5.0 // indirect
-	github.com/Masterminds/sprig v2.22.0+incompatible
 	github.com/google/uuid v1.3.0 // indirect
 	github.com/huandu/xstrings v1.3.2 // indirect
 	github.com/imdario/mergo v0.3.12 // indirect
 	github.com/mitchellh/copystructure v1.2.0 // indirect
+	github.com/mitchellh/reflectwalk v1.0.2 // indirect
 	github.com/stretchr/testify v1.7.1 // indirect
 	golang.org/x/crypto v0.0.0-20220411220226-7b82a4e95df4 // indirect
-	gopkg.in/yaml.v3 v3.0.0
 )
