@@ -26,7 +26,9 @@ describe('Site', () => {
     // Wait for results and check that there is at least one of them.
     await page.waitForSelector('div#results');
     await page.waitForSelector('li > div.summary');
-    await expect(page).toMatch('the basics of what neural networks are');
+    await expect(page).toMatchTextContent(
+      'the basics of what neural networks are'
+    );
   });
 
   it('should properly load blog assets', async () => {
@@ -34,7 +36,9 @@ describe('Site', () => {
     await page.evaluate(() => (window.location.pathname = '/blog/go-mlp/'));
     // Wait for main blog div to appear.
     await page.waitForSelector('div.type-blog');
-    await expect(page).toMatch('multi-layer perceptron from scratch');
+    await expect(page).toMatchTextContent(
+      'multi-layer perceptron from scratch'
+    );
     // Wait for mathjax rendering to appear.
     await page.waitForSelector('mjx-math');
     // Wait for emgithub to appear.
