@@ -5,7 +5,7 @@ module.exports = {
   // Specity launch parameters for puppeteer.
   launch: {
     dumpio: process.env.PUPPETEER_DUMPIO || false,
-    headless: process.env.HEADLESS !== 'false',
+    headless: process.env.HEADLESS !== 'false' ? 'new' : false,
     product: process.env.BROWSER || 'chrome',
   },
 };
